Removed redundancies, added GET support
A lot of code to handle errors was redundant inside the individual actions. This is now handled by anyCommonErrors(), which can not completely replace the individual actions because checkConfig() still has an additional error to catch. Also added a GET function, for use in collecting discovery_info and future use.